Import simplebgc directory from maiermic/robot-cameraman
Exact copy of that directory at b2c1c361b45952861784a03e71c7c19414b46f16
for tracking my changes to it.  That code was the first hit for python
SimpleBGC API code, outdated but does the job as a starting point.

Locally I added it after I had the UI but adding the simplebgc module
first because the UI now depends on the module.